refactor(observability): split SipiSentry into header + .cpp (DEV-6491)
Move all `inline` impls from `include/SipiSentry.h` (244 lines, fully
inline) into a new `src/SipiSentry.cpp`. The header previously pulled in
`SipiImage.hpp` (plus its transitive `metadata/SipiIcc.h`, exiv2, lcms2
chain) at every TU calling `capture_image_error`; consumers now see only
`<string>`, `<sentry.h>`, and the small `iiifparser/SipiQualityFormat.h`.

Public surface unchanged: `ImageContext`, `SipiMode`, `get_file_size`,
`format_type_to_string`, `populate_from_image`, `capture_image_error`.
The helpers `predefined_profile_to_string` and `orientation_to_string`
are demoted to file-static (anonymous namespace) since neither was called
externally.

`SipiImage` is now a forward declaration in the header; the full type is
brought in by `SipiSentry.cpp` alone. `<sys/stat.h>` likewise moves to
the .cpp. Both call sites (`src/sipi.cpp`, `src/SipiHttpServer.cpp`)
already include `SipiImage.hpp` directly, so no transitive-dep breakage.

Standalone compile-time speedup; first of three PRs in the Probe-9a
observability/ refactor. Sibling work: DEV-6492 + DEV-6493 establish the
new `src/observability/` directory; DEV-6496 promotes the Bazel package.

Refs DEV-6489.
